requirements.txt
requests
PyMuPDF
python-docx
pyyaml
openai
watchdog
beautifulsoup4
pandas
openpyxl
selenium
webdriver-manager
rich
psycopg2-binary
sqlalchemy
